requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.4
SQLAlchemy==2.0.39